Refactor and cleanup
