+++ bots/structured_log.py
@@ -1,8 +1,6 @@
 import os
-from supabase import create_client, Client
-from box import Box
 import dataclasses
-from typing import Any, Dict
+from supabase import create_client, Client
 from utils.slack_utils import SlackContext
 
 
@@ -12,6 +10,7 @@ class BotLogEntry:
     elapsed_ms: int
     step_name: str
     payload: object
+    # llm_rag_feedback: object
 
 # create single supabase client
 supabase: Client = create_client(os.environ['SLACK_BOT_SUPABASE_URL'],
@@ -19,11 +18,8 @@ supabase: Client = create_client(os.environ['SLACK_BOT_SUPABASE_URL'],
 
 
 def bot_log(entry: BotLogEntry):
-    # Convert the BotLogEntry object to a dictionary
-    # entry_dict = entry.__dict__
-    # Convert the dictionary to a JSON string
-    # entry_json = json.dumps(entry_dict)
-    insert_response = supabase.table('bot_log').insert(dataclasses.asdict(entry)).execute()
 
+    insert_response = supabase.table('bot_log').insert(dataclasses.asdict(entry)).execute()
+    
     print(f'insert_response: {insert_response}')
     return insert_response
